requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
scipy>=1.10.0
matplotlib>=3.7.0
supervision>=0.18.0